Updating HW data

+++ run_hw/get_hw_data.sh
@@ -1,4 +1,12 @@
-#!/usr/env sh
-wget https://engineering.purdue.edu/tgrogers/gpgpu-sim/hw_data/pascal.titanx.cycle.tgz .
-tar -xzvf pascal.titanx.cycle.tgz
-rm pascal.titanx.cycle.tgz
+export DATA_ROOT="$( cd "$( dirname "$BASH_SOURCE" )" && pwd )"
+TITANX_PASCAL_DATA=$DATA_ROOT/TITAN-X-PASCAL
+
+if [ ! -d $TITANX_PASCAL_DATA ]; then
+    echo "in"
+    wget https://engineering.purdue.edu/tgrogers/gpgpu-sim/hw_data/pascal.titanx.cycle.tgz .
+    tar -xzvf pascal.titanx.cycle.tgz -C $DATA_ROOT
+    rm pascal.titanx.cycle.tgz
+fi
+
+
+
